+++ modules/client/src/addressList/internal/client/methods.ts
@@ -2,11 +2,13 @@ import {
   GraphQLError,
   InvalidAddressError,
   InvalidAddressOrEnsError,
+  InvalidCidError,
   InvalidProposalIdError,
   IpfsPinError,
   NoProviderError,
   ProposalCreationError,
   Random,
+  resolveIpfsCid,
 } from "@aragon/sdk-common";
 import { isAddress } from "@ethersproject/address";
 import {
@@ -46,6 +48,10 @@ import { AllowlistVoting__factory } from "@aragon/core-contracts-ethers";
 import { id } from "@ethersproject/hash";
 import { hexZeroPad } from "@ethersproject/bytes";
 import { toUtf8Bytes } from "@ethersproject/strings";
+import {
+  UNAVAILABLE_PROPOSAL_METADATA,
+  UNSUPPORTED_PROPOSAL_METADATA_LINK,
+} from "../../../client-common/constants";
 
 /**
  * Methods module the SDK Address List Client
@@ -275,11 +281,24 @@ export class ClientAddressListMethods extends ClientCore
       if (!addressListProposal) {
         return null;
       }
-      // format in the metadata field
-      const metadataString = await this.ipfs.fetchString(addressListProposal.metadata);
-      // TODO: Parse and validate schema
-      const metadata = JSON.parse(metadataString) as ProposalMetadata;
-      return toAddressListProposal(addressListProposal, metadata);
+      try {
+        const metadataCid = resolveIpfsCid(addressListProposal.metadata);
+        const metadataString = await this.ipfs.fetchString(metadataCid);
+        const metadata = JSON.parse(metadataString) as ProposalMetadata;
+        return toAddressListProposal(addressListProposal, metadata);
+        // TODO: Parse and validate schema
+      } catch (err) {
+        if (err instanceof InvalidCidError) {
+          return toAddressListProposal(
+            addressListProposal,
+            UNSUPPORTED_PROPOSAL_METADATA_LINK,
+          );
+        }
+        return toAddressListProposal(
+          addressListProposal,
+          UNAVAILABLE_PROPOSAL_METADATA,
+        );
+      }
     } catch (err) {
       throw new GraphQLError("AddressList proposal");
     }
@@ -344,17 +363,27 @@ export class ClientAddressListMethods extends ClientCore
       await this.ipfs.ensureOnline();
       return Promise.all(
         addressListProposals.map(
-          (
+          async (
             proposal: SubgraphAddressListProposalListItem,
           ): Promise<AddressListProposalListItem> => {
             // format in the metadata field
-            return this.ipfs
-              .fetchString(proposal.metadata)
-              .then((stringMetadata: string) => {
-                // TODO: Parse and validate schema
-                const metadata = JSON.parse(stringMetadata) as ProposalMetadata;
-                return toAddressListProposalListItem(proposal, metadata);
-              });
+            try {
+              const metadataCid = resolveIpfsCid(proposal.metadata);
+              const stringMetadata = await this.ipfs.fetchString(metadataCid);
+              const metadata = JSON.parse(stringMetadata) as ProposalMetadata;
+              return toAddressListProposalListItem(proposal, metadata);
+            } catch (err) {
+              if (err instanceof InvalidCidError) {
+                return toAddressListProposalListItem(
+                  proposal,
+                  UNSUPPORTED_PROPOSAL_METADATA_LINK,
+                );
+              }
+              return toAddressListProposalListItem(
+                proposal,
+                UNAVAILABLE_PROPOSAL_METADATA,
+              );
+            }
           },
         ),
       );