+++ modules/client/src/internal/client/methods.ts
@@ -8,8 +8,10 @@ import {
 import {
   GraphQLError,
   InvalidAddressOrEnsError,
+  InvalidCidError,
   NoProviderError,
   NoSignerError,
+  resolveIpfsCid,
 } from "@aragon/sdk-common";
 import { Signer } from "@ethersproject/abstract-signer";
 import { BigNumber } from "@ethersproject/bignumber";
@@ -40,6 +42,7 @@ import {
   IDaoQueryParams,
   IDepositParams,
   IHasPermissionParams,
+  IMetadata,
   ITransferQueryParams,
   SubgraphBalance,
   SubgraphDao,
@@ -49,11 +52,7 @@ import {
   Transfer,
   TransferSortBy,
 } from "../../interfaces";
-import {
-  ClientCore,
-  Context,
-  SortDirection,
-} from "../../client-common";
+import { ClientCore, Context, SortDirection } from "../../client-common";
 import {
   toAssetBalance,
   toDaoDetails,
@@ -64,6 +63,10 @@ import {
 import { isAddress } from "@ethersproject/address";
 import { toUtf8Bytes, toUtf8String } from "@ethersproject/strings";
 import { id } from "@ethersproject/hash";
+import {
+  UNAVAILABLE_DAO_METADATA,
+  UNSUPPORTED_DAO_METADATA_LINK,
+} from "../constants";
 
 /**
  * Methods module the SDK Generic Client
@@ -267,7 +270,7 @@ export class ClientMethods extends ClientCore implements IClientMethods {
       if (!log) {
         throw new Error("Could not find the Approval event");
       }
-      const value = log.data
+      const value = log.data;
       if (!value || BigNumber.from(amount).gt(BigNumber.from(value))) {
         throw new Error("Could not update allowance");
       }
@@ -336,11 +339,18 @@ export class ClientMethods extends ClientCore implements IClientMethods {
       if (!dao) {
         return null;
       }
-      const stringMetadata = await this.ipfs.fetchString(
-        dao.metadata,
-      );
-      const metadata = JSON.parse(stringMetadata);
-      return toDaoDetails(dao, metadata);
+      try {
+        const metadataCid = resolveIpfsCid(dao.metadata);
+        const metadataString = await this.ipfs.fetchString(metadataCid);
+        const metadata = JSON.parse(metadataString) as IMetadata;
+        return toDaoDetails(dao, metadata);
+        // TODO: Parse and validate schema
+      } catch (err) {
+        if (err instanceof InvalidCidError) {
+          return toDaoDetails(dao, UNSUPPORTED_DAO_METADATA_LINK);
+        }
+        return toDaoDetails(dao, UNAVAILABLE_DAO_METADATA);
+      }
     } catch (err) {
       throw new GraphQLError("DAO");
     }
@@ -378,11 +388,18 @@ export class ClientMethods extends ClientCore implements IClientMethods {
       await this.ipfs.ensureOnline();
       return Promise.all(
         daos.map(
-          (dao: SubgraphDaoListItem): Promise<DaoListItem> => {
-            return this.ipfs.fetchString(dao.metadata).then((stringMetadata) => {
+          async (dao: SubgraphDaoListItem): Promise<DaoListItem> => {
+            try {
+              const metadataCid = resolveIpfsCid(dao.metadata);
+              const stringMetadata = await this.ipfs.fetchString(metadataCid);
               const metadata = JSON.parse(stringMetadata);
               return toDaoListItem(dao, metadata);
-            });
+            } catch (err) {
+              if (err instanceof InvalidCidError) {
+                return toDaoListItem(dao, UNSUPPORTED_DAO_METADATA_LINK);
+              }
+              return toDaoListItem(dao, UNAVAILABLE_DAO_METADATA);
+            }
           },
         ),
       );