+++ modules/common/src/index.ts
@@ -2,6 +2,7 @@ export * from "./constants";
 export * from "./encoding";
 export * from "./errors";
 export * from "./promises";
+export * from "./uri";
 export * from "./random";
 export * from "./timeout";
 export * from "./types";