+++ modules/client/CHANGELOG.md
@@ -17,6 +17,9 @@ TEMPLATE:
 -->
 
 ## [UPCOMING]
+### Fixed
+- Reading metadata back now parses the IPFS URI schema
+
 
 ## 0.15.1-alpha
 On 2022-12-06 11:20:50